feat: add timedelta support (#134)
As requested in #133, I have added support for parsing and converting an
ISO 8601 duration to a `timedelta`. Additionally, I have added automated
tests to verify the behavior and a usage example in the `README.md`.
Please feel free to let me know if I failed to add tests somewhere that
should really include some tests.